# packages/remote-contracts/references/backend-openapi.yaml
openapi: 3.0.0
info:
  title: README PDF Reader API
  description: README的OpenAPI文档
  version: 1.2.0
servers:
  - url: http://localhost:5000/api
    description: 本地开发环境
security:
  - BearerAuth: []
paths:
...
  /notes:
    post:
      tags:
        - Notes
      summary: 创建笔记
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteActionResponse'
        '400':
          description: 缺少 pdfId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /notes/{pdf_id}:
    get:
      tags:
        - Notes
      summary: 获取某 PDF 的所有笔记
      security:
        - BearerAuth: []
      description: 'Flask 路由: GET /notes/<pdf_id>'
      parameters:
        - name: pdf_id
          in: path
          required: true
          description: PDF 文件 Hash
          schema:
            type: string
      responses:
        '200':
          description: 笔记列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteListResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /notes/{note_id}:
    put:
      tags:
        - Notes
      summary: 更新笔记
      security:
        - BearerAuth: []
      description: 'Flask 路由: PUT /notes/<int:note_id>'
      parameters:
        - name: note_id
          in: path
          required: true
          description: 笔记ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteActionResponse'
        '400':
          description: 未提供任何可更新字段
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Notes
      summary: 删除笔记
      security:
        - BearerAuth: []
      description: 'Flask 路由: DELETE /notes/<int:note_id>'
      parameters:
        - name: note_id
          in: path
          required: true
          description: 笔记ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteActionResponse'
        '404':
          description: 笔记不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /highlight:
    post:
      tags:
        - Highlight
      summary: 创建高亮
      security:
        - BearerAuth: []
      description: 前端画了高亮以后，后端要存到数据库。这里的传入坐标是像素坐标。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHighlightRequest'
      responses:
        '200':
          description: 创建成功，返回归一化后的坐标，也就是0.0-1.0的坐标
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateHighlightResponse'
        '400':
          description: 缺少必填字段
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: PDF 不在用户库中
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Highlight
      summary: 获取某 PDF 的所有高亮
      security:
        - BearerAuth: []
      description: 返回归一化坐标 (0.0-1.0)，前端需乘以当前 Canvas 尺寸渲染
      parameters:
        - name: pdfId
          in: query
          required: true
          description: PDF 文件 Hash
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: 按页码筛选
          schema:
            type: integer
      responses:
        '200':
          description: 高亮列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HighlightListResponse'
        '400':
          description: 缺少 pdfId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /highlight/{highlight_id}:
    put:
      tags:
        - Highlight
      summary: 更新高亮颜色
      security:
        - BearerAuth: []
      parameters:
        - name: highlight_id
          in: path
          required: true
          description: 高亮ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateHighlightRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Highlight
      summary: 删除高亮
      security:
        - BearerAuth: []
      parameters:
        - name: highlight_id
          in: path
          required: true
          description: 高亮ID
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /translate/paragraph:
    post:
      tags:
        - Translate
      summary: 翻译段落（按需触发，带缓存）
      security:
        - BearerAuth: []
      description: |
        前端点击某段落的翻译按钮时调用。
        后端流程：解析段落 ID → 从 DB 取原文 → 调用 LLM 翻译 → 结果存库。
        已翻译过的段落会命中缓存直接返回，除非 force=true。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslateParagraphRequest'
      responses:
        '200':
          description: 翻译结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateParagraphResponse'
        '400':
          description: 缺少参数或段落 ID 格式错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 段落原文为空
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /translate/text:
    post:
      tags:
        - Translate
      summary: 翻译选中文本（带上下文）
      security:
        - BearerAuth: []
      description: |
        用户框选一段文字后调用。
        如果提供了 pdfId，后端会取前 N 段原文作为上下文，帮助 LLM 更准确地翻译。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranslateTextRequest'
      responses:
        '200':
          description: 翻译结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranslateTextResponse'
        '400':
          description: 缺少 text 或 text 为空
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /kg/nodes/{node_id}/notes:
    post:
      tags:
        - KnowledgeGraph
      summary: 关联笔记到节点
      description: 将笔记关联到指定节点
      security:
        - BearerAuth: []
      parameters:
        - name: node_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkNoteRequest'
      responses:
        '200':
          description: 关联成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Note linked to node successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 节点或笔记不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /kg/nodes/{node_id}/notes/{note_id}:
    delete:
      tags:
        - KnowledgeGraph
      summary: 取消笔记关联
      description: 取消笔记与节点的关联
      security:
        - BearerAuth: []
      parameters:
        - name: node_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: note_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 取消关联成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Note unlinked from node successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 节点或笔记不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Access Token，通过 /api/auth/login 或 /api/auth/register 获取。
        在请求头中以 `Authorization: Bearer <token>` 方式携带。
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        code:
          type: string
          description: 机器可读的错误码，前端按此字段判断错误类型
          example: MISSING_FIELDS
        error:
          type: string
          description: 人类可读的错误信息，可直接展示给用户
          example: Invalid file type
...
    CreateNoteRequest:
      type: object
      required:
        - pdfId
      properties:
        pdfId:
          type: string
          description: 关联的 PDF 文件 Hash
        title:
          type: string
          description: 笔记标题（可选）
        content:
          type: string
          description: 笔记内容，Markdown 或 JSON（可选）
        keywords:
          type: array
          items:
            type: string
          description: 关键词列表（可选）
    NoteActionResponse:
      type: object
      description: 创建/更新/删除笔记的通用成功响应
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        id:
          type: integer
          description: 笔记ID，仅创建时返回，更新和删除时为空
        message:
          type: string
          example: Note created
    Note:
      type: object
      properties:
        id:
          type: integer
          description: 笔记ID
        title:
          type: string
          description: 笔记标题
          example: Transformer 架构要点
        content:
          type: string
          description: 笔记内容（Markdown 或 JSON）
          example: |-
            ## 核心思想
            Self-Attention 机制...
        keywords:
          type: array
          items:
            type: string
          description: 关键词列表
          example:
            - transformer
            - attention
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    NoteListResponse:
      type: object
      properties:
        success:
          type: boolean
        notes:
          type: array
          items:
            $ref: '#/components/schemas/Note'
        total:
          type: integer
    UpdateNoteRequest:
      type: object
      description: 至少需要提供 title、content、keywords 中的一个字段，也就是至少更新一个
      properties:
        title:
          type: string
        content:
          type: string
        keywords:
          type: array
          items:
            type: string
    Rect:
      type: object
      description: 归一化后的矩形坐标 (0.0 - 1.0)，前端需乘以当前 Canvas 尺寸渲染
      required:
        - x
        - 'y'
        - width
        - height
      properties:
        x:
          type: number
          format: float
        'y':
          type: number
          format: float
        width:
          type: number
          format: float
        height:
          type: number
          format: float
    Highlight:
      type: object
      properties:
        id:
          type: integer
        page:
          type: integer
        rects:
          type: array
          items:
            $ref: '#/components/schemas/Rect'
        text:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
    HighlightListResponse:
      type: object
      properties:
        success:
          type: boolean
        highlights:
          type: array
          items:
            $ref: '#/components/schemas/Highlight'
        total:
          type: integer
    CreateHighlightRequest:
      type: object
      required:
        - pdfId
        - page
        - rects
        - pageWidth
        - pageHeight
        - text
        - color
      description: 创建高亮请求。text和color待定是否可选
      properties:
        pdfId:
          type: string
          description: PDF 文件 Hash
        page:
          type: integer
          description: 页码
          example: 1
        rects:
          type: array
          description: 选区矩形列表，是像素坐标
          items:
            $ref: '#/components/schemas/Rect'
        pageWidth:
          type: number
          description: 当前页面宽度（像素），用于坐标归一化
          example: 800
        pageHeight:
          type: number
          description: 当前页面高度（像素），用于坐标归一化
          example: 1200
        text:
          type: string
          description: 选中的文本内容
        color:
          type: string
          description: 高亮颜色，格式是16进制
          default: '#FFFF00'
          example: '#FFFF00'
    CreateHighlightResponse:
      type: object
      properties:
        success:
          type: boolean
        id:
          type: integer
          description: 新建高亮ID
        rects:
          type: array
          description: 归一化后的矩形坐标
          items:
            $ref: '#/components/schemas/Rect'
        message:
          type: string
          example: Highlight created
    UpdateHighlightRequest:
      type: object
      properties:
        color:
          type: string
          description: 新颜色
          example: '#FF0000'
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
    TranslateParagraphRequest:
      type: object
      required:
        - pdfId
        - paragraphId
      properties:
        pdfId:
          type: string
          description: PDF 文件 Hash
        paragraphId:
          type: string
          description: 段落 ID（格式：pdf_chk_{hash前8位}_{页码}_{块号}）
          example: pdf_chk_11451419_3_2
        force:
          type: boolean
          description: 是否强制重新翻译（忽略缓存）
          default: false
        model:
          type: string
          description: 指定使用的模型（覆盖全局配置）
        apiKey:
          type: string
          description: 指定使用的 API Key
        apiBase:
          type: string
          description: 指定使用的 API Base URL
    TranslateParagraphResponse:
      type: object
      properties:
        success:
          type: boolean
        translation:
          type: string
          description: 翻译结果
        cached:
          type: boolean
          description: 是否命中缓存
        paragraphId:
          type: string
    TranslateTextRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: 选中的文本
          example: Self-attention mechanism
        pdfId:
          type: string
          description: 关联的 PDF Hash（可选，用于获取上下文提高翻译质量）
        contextParagraphs:
          type: integer
          description: 用于上下文的段落数量
          default: 3
        model:
          type: string
          description: 指定使用的模型（覆盖全局配置）
        apiKey:
          type: string
          description: 指定使用的 API Key
        apiBase:
          type: string
          description: 指定使用的 API Base URL
    TranslateTextResponse:
      type: object
      properties:
        success:
          type: boolean
        originalText:
          type: string
          description: 原文
        translatedText:
          type: string
          description: 翻译结果
        hasContext:
          type: boolean
          description: 是否使用了上下文
        contextLength:
          type: integer
          description: 上下文长度
...
    LinkNoteRequest:
      type: object
      required:
        - note_id
      properties:
        note_id:
          type: integer
          description: 笔记ID
...
  responses:
    Unauthorized:
      description: 未认证或 token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...